requirements.txt
asyncpg
pytest~=7.2.0
requests~=2.28.1
sqlalchemy~=1.4.44
psycopg2-binary~=2.9
python-dotenv~=1.0.0
aiohttp[speedups]~=3.8.4
